+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* Autocorrect of method order would not replace all incorrectly ordered nodes in
+  certain combinations.
+
 ## [0.1.0] - 2018-03-21
 
 ### Added